Turned Map into less reading
