+++ eat_all_fungus/lib/controllers/tileMapController.dart
@@ -5,18 +5,18 @@ import 'package:eat_all_fungus/providers/streams/playerStream.dart';
 import 'package:eat_all_fungus/services/tileRepository.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
-final tileMapControllerProvider = StateNotifierProvider<TileMapController,
-    AsyncValue<Map<int, Map<int, MapTile>>>>((ref) {
+final tileMapControllerProvider =
+    StateNotifierProvider<TileMapController, Map<int, Map<int, MapTile>>?>(
+        (ref) {
   final player = ref.watch(playerStreamProvider);
   return TileMapController(ref.read, player);
 });
 
-class TileMapController
-    extends StateNotifier<AsyncValue<Map<int, Map<int, MapTile>>>> {
+class TileMapController extends StateNotifier<Map<int, Map<int, MapTile>>?> {
   final Reader _read;
   final Player? _player;
 
-  TileMapController(this._read, this._player) : super(AsyncValue.loading()) {
+  TileMapController(this._read, this._player) : super(null) {
     if (_player != null) {
       getCurrentTileMap();
     }
@@ -27,10 +27,11 @@ class TileMapController
       final tileMap = await _read(mapTileRepository)
           .getTilesInWorld(worldID: _player!.worldID);
       if (mounted) {
-        state = AsyncValue.data(tileMap);
+        state = tileMap;
       }
-    } on CustomException catch (error, stackTrace) {
-      state = AsyncValue.error(error, stackTrace);
+    } on CustomException catch (error) {
+      print('TileMapController - ${error.message}');
+      state = null;
     }
   }
 }
